feat: Add fcmToken logging in PermissionScreen and handleNotificationAction
This commit adds logging statements to the PermissionScreen component to log the fcmToken when it changes. It also adds logging statements to the handleNotificationAction function in the notifee.ts file to log the details of the notification action. These logging statements will help with debugging and monitoring the behavior of the notification system.