feat: Add APP_NAME constant and update notification handling
This commit adds the `APP_NAME` constant to the `env.config.ts` file, setting it to 'Middo Extension'. It also updates the notification handling in the `fcm.ts` file to check if the notification is from the extension and should be displayed. This improves the notification system by filtering out notifications that are not relevant to the current room and are not from the extension.